+++ crates/voicevox_core_java_api/lib/src/main/java/jp/hiroshiba/voicevoxcore/exceptions/GetSupportedDevicesException.java
@@ -3,7 +3,7 @@ package jp.hiroshiba.voicevoxcore.exceptions;
 import java.io.IOException;
 
 /** サポートされているデバイス情報取得に失敗した。 */
-public class GetSupportedDevicesException extends IOException {
+public final class GetSupportedDevicesException extends IOException {
   public GetSupportedDevicesException(String message) {
     super(message);
   }