+++ crates/voicevox_core_java_api/lib/src/main/java/jp/hiroshiba/voicevoxcore/exceptions/InvalidWordException.java
@@ -1,7 +1,7 @@
 package jp.hiroshiba.voicevoxcore.exceptions;
 
 /** ユーザー辞書の単語のバリデーションに失敗した。 */
-public class InvalidWordException extends IllegalArgumentException {
+public final class InvalidWordException extends IllegalArgumentException {
   public InvalidWordException(String message) {
     super(message);
   }