+++ crates/voicevox_core_java_api/lib/src/main/java/jp/hiroshiba/voicevoxcore/exceptions/StyleNotFoundException.java
@@ -1,7 +1,7 @@
 package jp.hiroshiba.voicevoxcore.exceptions;
 
 /** スタイルIDに対するスタイルが見つからなかった。 */
-public class StyleNotFoundException extends IndexOutOfBoundsException {
+public final class StyleNotFoundException extends IndexOutOfBoundsException {
   public StyleNotFoundException(String message) {
     super(message);
   }